Pass vector of S3 URIs to arrow::open_dataset
This is a test to see if passing a list of S3 URIs to open_dataset
would allow us to handle mixed model-output file types AND still
get performance benefits. The theory being that telling open_dataset
exactly which files to include would be faster than using the
exclude_invalid_files = TRUE option.